pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weld-traceability-assistant"
version = "0.1.0"
description = "Modular pipeline for weld drawing understanding and weld traceability."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
  "Pillow>=10.0",
  "numpy>=2.0",
  "pydantic>=2.9",
  "ollama>=0.4.0"
]

[project.optional-dependencies]
ui = ["streamlit>=1.40"]
ocr = ["paddleocr>=2.7"]
validation = ["jsonschema>=4.23"]
dev = ["pytest>=8.3", "streamlit>=1.40", "jsonschema>=4.23"]

[project.scripts]
weld-assistant = "weld_assistant.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

